+++ sdk/communication/azure-communication-callingserver/src/main/java/module-info.java
@@ -13,4 +13,5 @@ module com.azure.communication.callingserver {
     // exporting some packages specifically for Jackson
     opens com.azure.communication.callingserver.models to com.fasterxml.jackson.databind;
     opens com.azure.communication.callingserver.implementation.models to com.fasterxml.jackson.databind, com.azure.core;
+    opens com.azure.communication.callingserver to com.fasterxml.jackson.databind;
 }